Started with a form and broke it into pieces
Here is the start of my simple login assignment. I started with the
main.py file included with the google launcher and added my html for my
form and broke it up between page_head, page_body and page_close so
that I can write the write the form when I want to and leave it out
when I don’t want to write the form. I added six elements to the form:
first name, last name, email, phone, preferred contact method as a
radio button and finally a reason for contact as a select method.