Fixed errors preventing code to run properly
My original code did not work because of some indentation errors and
html errors. I fixed the indentation errors using the logs in google
launcher and the html errors I made a separate html file with the html
code and ran it through an html validator found the errors and
corrected my code. I had misspelled label in a few spots and forgot a
couple tags.