Functional application changed the input information.
I finally got the application functional. I was putting the instance of
the class in the wrong places. I put it inside the if statement and it
works. I changed the application to make more sense for this assignment
by collecting information about a customer’s address information for a
pizza delivery.