Finished application
Forgot to mention in previous commit that a pages.py was created to
ensure that the html comes from a class in a different files. I also
imported it into the application as well as the css. I also added the
css handler in the yaml file. Finally I added a few comments to the
code.