Setup for the reusable library assignment
I have set up my main handler to import the Page class from the page
file. I set the variable p to the page class and then printed out the
page using the p.printout() function from the page class. I set up my
page file with a class called page to handle the majority of my html
separating them into sections head, body and close. I set up a basic
form with the get method, 5 inputs and a submit button for the form and
wrote the method to use to print out the html. I also made a library
file but have not added anything yet. All comments so far are some of
my to-do lists for what still needs to be done.