added bulk of code
Here I programmed the bulk of my code. I finished the page.py file with
all of my html elements and a two different print out functions one for
the form view and one for the results view. The main.py imports both of
the Page class from page and the HighScores data object class and the
FavoriteGame reusable library class I set up an instance for Page() and
an instance of FavoriteGame. For the form I added an instance of the
data object and put in a conditional statement so that if the form is
filled out it shows the results view and if it is not it shows the form
view. In the library file I have my data object HighScores with a
private attribute of score and I set up a getter and setter for the
attribute. I also created the FavoriteGame library with a method that
creates an array, a method that adds a game and method that compiles a
list of added games and outputs the information. I also added css to
the yaml file and added a css file for styling