dynamic site set up
I started my dynamic site by creating a basic html structure in a page
class and created it in a page.py file and then I created two data
objects one for shirts and one for albums and then put the data in.
The data is is still in progress but it is a start.