work in progress
So i have done a lot of work I renamed some variables in the data file
and changed it so that they are accessible through an array. I have
added some else if statements in the main handler but they still need
work I am still trying to figure out how to use the data from the data
object in the page without putting the data object in the page file.